[quantization] Use ConvPackedParams everywhere
[ghstack-poisoned]
