+++ torch/nn/intrinsic/quantized/modules/conv_relu.py
@@ -34,7 +34,7 @@ class ConvReLU2d(nnq.Conv2d):
         if len(input.shape) != 4:
             raise ValueError("Input shape must be `(N, C, H, W)`!")
         return torch.ops.quantized.conv2d_relu(
-            input, self._packed_params, self.stride, self.padding,
+            input, self._packed_params._packed_params, self.stride, self.padding,
             self.dilation, self.groups, self.scale, self.zero_point)
 
     def _get_name(self):
@@ -75,7 +75,7 @@ class ConvReLU3d(nnq.Conv3d):
         if len(input.shape) != 5:
             raise ValueError("Input shape must be `(N, C, D, H, W)`!")
         return torch.ops.quantized.conv3d_relu(
-            input, self._packed_params, self.stride, self.padding,
+            input, self._packed_params._packed_params, self.stride, self.padding,
             self.dilation, self.groups, self.scale, self.zero_point)
 
     def _get_name(self):
