+++ torch/quantization/_quantize_script.py
@@ -3,62 +3,11 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 import torch
 from .qconfig import QConfig
 
-class ConvPackedParams(torch.nn.Module):
-    def __init__(self):
-        super(ConvPackedParams, self).__init__()
-        wq = torch._empty_affine_quantized([1, 1, 1, 1], scale=1.0, zero_point=0, dtype=torch.qint8)
-        self.stride = [1, 1]
-        self.padding = [0, 0]
-        self.dilation = [1, 1]
-        self.groups = 1
-        self.set_weight_bias(wq, None)
-
-    @torch.jit.export
-    def set_conv_params(self, stride, padding, dilation, groups):
-        # type: (List[int], List[int], List[int], int) -> None
-        self.stride = stride
-        self.padding = padding
-        self.dilation = dilation
-        self.groups = groups
-
-    @torch.jit.export
-    def set_weight_bias(self, weight, bias):
-        # type: (torch.Tensor, Optional[torch.Tensor]) -> None
-        self._packed_params = torch.ops.quantized.conv2d_prepack(weight, bias, self.stride, self.padding, self.dilation, self.groups)
-
-    @torch.jit.export
-    def _weight_bias(self):
-        return torch.ops.quantized.conv2d_unpack(self._packed_params)
-
-    def forward(self, x):
-        return x
-
-    @torch.jit.export
-    def __getstate__(self):
-        qweight, bias = self._weight_bias()
-        return (qweight,
-                bias,
-                self.stride,
-                self.padding,
-                self.dilation,
-                self.groups,
-                self.training)
-
-    @torch.jit.export
-    def __setstate__(self, state):
-        self.stride = state[2]
-        self.padding = state[3]
-        self.dilation = state[4]
-        self.groups = state[5]
-        self.set_weight_bias(state[0],
-                             state[1])
-        self.training = state[6]
-
 linear_packed_params = None
 conv_packed_params = None
 if 'fbgemm' in torch.backends.quantized.supported_engines:
     linear_packed_params = torch.jit.script(torch.nn.quantized.modules.linear.LinearPackedParams())._c
-    conv_packed_params = torch.jit.script(ConvPackedParams())._c
+    conv_packed_params = torch.jit.script(torch.nn.quantized.modules.conv.ConvPackedParams())._c
 
 def _check_is_script_module(model):
     if not isinstance(model, torch.jit.ScriptModule):
